order-inventory-observability/services/inventory/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2

opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-prometheus==0.50b0
prometheus-client==0.21.0